created single reducer
