+++ src/components/CakeContainer.js
@@ -15,7 +15,7 @@ function CakeContainer(props) {
 
 const mapStateToProps = state => {
     return{
-        numOfCakes:state.cake.numOfCakes
+        numOfCakes:state.numOfCakes
      }
 }
