+++ src/redux/store.js
@@ -1,9 +1,9 @@
 import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
-import rootReducer from '../redux/rootReducer'
+import Reducer from './reducer/Reducer'
 import logger from 'redux-logger'
 
-const store= createStore(rootReducer, composeWithDevTools(applyMiddleware(logger, thunk)) )
+const store= createStore(Reducer, composeWithDevTools(applyMiddleware(logger, thunk)) )
 
 export default store
